refactoring s@k
